Added support for large ints and did day 9

+++ src/day5/day5_test.go
@@ -6,11 +6,11 @@ import (
 )
 
 func TestProgram(t *testing.T) {
-	var testProgramB intcomp.Program = intcomp.CreateProgram([]int{3, 3, 1107, -1, 8, 3, 4, 3, 99})
+	var testProgramB intcomp.Program = intcomp.CreateProgram([]int64{3, 3, 1107, -1, 8, 3, 4, 3, 99})
 	go testProgramB.ProcessProgram()
 	testProgramB.Input <- 1
 	result := <-testProgramB.Output
-	expected := 1
+	expected := int64(1)
 	if result != expected {
 		t.Errorf("Got result %v, expected %v", result, expected)
 	}
